requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argcomplete==3.6.2
asgiref==3.9.1
attrs==25.3.0
audioread==3.0.1
av==15.0.0
beautifulsoup4==4.13.4
black==25.1.0
certifi==2025.7.14
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.1
click-default-group==1.2.4
coloredlogs==15.0.1
ctranslate2==4.6.0
datamodel-code-generator==0.26.1
datasets==4.0.0
decorator==5.2.1
dill==0.3.8
Django==5.2.4
djangorestframework==3.16.0
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
faiss-cpu==1.11.0.post1
Faker==37.4.2
faster-whisper==1.1.1
filelock==3.18.0
flake8==7.3.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.3.0
genson==1.3.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.10
ijson==3.4.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
inflect==5.6.2
isort==5.13.2
Jinja2==3.1.6
joblib==1.5.1
jsf==0.11.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
label-studio-sdk==1.0.19
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
lxml==6.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnxruntime==1.22.1
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
proglog==0.1.12
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyarrow==21.0.0
PyAudio==0.2.14
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pyflakes==3.4.0
pyjsparser==2.7.1
PyJWT==2.10.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-mock==1.12.1
rpds-py==0.26.0
rstr==3.2.2
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
sentence-transformers==5.0.0
seqeval==1.2.2
setuptools==80.9.0
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
sqlite-fts4==1.0.3
sqlite-utils==3.38
sqlparse==0.5.3
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
torchaudio==2.7.1
tqdm==4.67.1
transformers==4.53.2
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.5.0
wrapt==1.17.2
xmljson==0.2.1
xxhash==3.5.0
yarl==1.20.1
